Move tests to mocha dir.
