--- test/vc.spec.js
@@ -1,19 +1,0 @@
-/*!
- * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
- */
-import * as vc from '@digitalbazaar/vc';
-import chai from 'chai';
-import {driver} from '../lib/index.js';
-
-chai.should();
-const {expect} = chai;
-
-describe('vc', function() {
-  describe('Sign', function() {
-
-  });
-  describe('Verify', function() {
-
-  });
-});
-
